revisited static_cast<> for integral types

+++ tom/src/tom/terminal.cpp
@@ -57,14 +57,14 @@ Terminal::TerminalSize Terminal::terminalSize() noexcept
     CONSOLE_SCREEN_BUFFER_INFO csbi;
     GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi);
 
-    width  = static_cast<int>(csbi.srWindow.Right - csbi.srWindow.Left) + 1;
-    height = static_cast<int>(csbi.srWindow.Bottom - csbi.srWindow.Top) + 1;
+    width  = (csbi.srWindow.Right - csbi.srWindow.Left) + 1;
+    height = (csbi.srWindow.Bottom - csbi.srWindow.Top) + 1;
 #elif defined(__linux__)
     winsize w {};
     ioctl(fileno(stdout), TIOCGWINSZ, &w);
 
-    width  = static_cast<int>(w.ws_col);
-    height = static_cast<int>(w.ws_row);
+    width  = w.ws_col;
+    height = w.ws_row;
 #endif
 
     return {width, height};
